Fix assumptions that CMAKE_INSTALL_*DIR variables are relative
The CMAKE_INSTALL_*DIR variables may be absolute, in which case they must not
be concatenated with CMAKE_INSTALL_PREFIX. If an absolute path is needed,
CMAKE_INSTALL_FULL_*DIR must be used. In other cases, more complex logic may be
required to make packages relocatable in the case that CMAKE_INSTALL_*DIR is
relative.

This patch replaces the custom LIB_INSTALL_DIR and BIN_INSTALL_DIR variables
with the standard GNUInstallDirs variables throughout the code. The custom
variables are still honored if specified manually.

Signed-off-by: Ben Wolsieffer <benwolsieffer@gmail.com>